requirements.txt
cassandra-driver==3.7.0
numpy==1.11.1
matplotlib=1.5.3

